Refactor Client out of ClientActor for easier testing
