+++ chain/client/src/lib.rs
@@ -1,4 +1,4 @@
-pub use crate::client::ClientActor;
+pub use crate::client_actor::ClientActor;
 pub use crate::types::{
     BlockProducer, ClientConfig, Error, GetBlock, GetChunk, Query, Status, StatusResponse,
     SyncStatus, TxDetails, TxStatus,
@@ -6,6 +6,7 @@ pub use crate::types::{
 pub use crate::view_client::ViewClientActor;
 
 mod client;
+mod client_actor;
 mod info;
 mod sync;
 pub mod test_utils;
